+++ src/subnet/gateway/routes/v1/miners.py
@@ -3,7 +3,7 @@ from fastapi import Depends, APIRouter
 from pydantic import BaseModel
 
 from src.subnet.validator.validator import Validator
-from src.subnet.validator_api import get_validator, api_key_auth
+from src.subnet.gateway import get_validator, api_key_auth
 
 miner_router = APIRouter(prefix="/v1/miner", tags=["miner"])