+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 ## [1.2.0] - 2024-10-30
 
 ### Added
-- Added commune ai-related  Validator API endpoints.
+- Added commune ai-related Validator API endpoints.
 - Add endpoint for returning available miners per network.